requirements.txt
pygame
pynput
audioop-lts